Combined config.cpp
